Update configuration to use Stripes CLI
