+++ .npmrc
@@ -1,1 +1,1 @@
-@folio:registry=https://repository.folio.org/repository/npm-folio/
+@folio:registry=https://repository.folio.org/repository/npm-folioci/
