+++ package.json
@@ -6,38 +6,26 @@
   "repository": "https://github.com/folio-org/ui-eholdings",
   "license": "Apache-2.0",
   "scripts": {
-    "postinstall": "bin/link-self",
-    "start": "NODE_ENV=development node demo/stripes.js dev demo/stripes.config.js",
-    "build": "node demo/stripes.js build demo/stripes.config.js dist",
-    "test": "karma start",
+    "start": "NODE_ENV=development stripescli serve demo/stripes.config.js",
+    "build": "stripescli build demo/stripes.config.js dist",
+    "test": "stripescli test demo/stripes.config.js --type=unit",
     "lint": "eslint ./"
   },
   "devDependencies": {
+    "@folio/stripes-cli": "folio-org/stripes-cli",
     "@folio/stripes-components": "folio-org/stripes-components#master",
     "@folio/stripes-core": "folio-org/stripes-core#master",
-    "babel-core": "^6.25.0",
-    "babel-eslint": "^8.0.1",
-    "babel-loader": "^7.1.1",
     "babel-plugin-istanbul": "^4.1.5",
-    "babel-preset-env": "^1.6.0",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-3": "^6.24.1",
     "chai": "^4.0.2",
     "chai-jquery": "^2.0.0",
     "eslint": "^4.8.0",
     "eslint-config-stripes": "^1.0.0",
     "eslint-loader": "^1.9.0",
     "jquery": "^3.2.1",
-    "karma": "^1.7.0",
     "karma-browserstack-launcher": "^1.3.0",
-    "karma-chrome-launcher": "^2.2.0",
     "karma-coverage": "^1.1.1",
-    "karma-mocha": "^1.3.0",
-    "karma-mocha-reporter": "^2.2.3",
-    "karma-webpack": "^2.0.3",
     "lodash": "^4.17.4",
     "mirage-server": "cowboyd/mirage-server",
-    "mocha": "^3.4.2",
     "moment": "^2.17.1",
     "react": "^15.6.1",
     "react-dom": "^15.6.1",
@@ -45,8 +33,7 @@
     "react-trigger-change": "^1.0.2",
     "stylelint": "^8.2.0",
     "stylelint-config-standard": "^17.0.0",
-    "webpack": "^3.0.0",
-    "webpack-dev-server": "^2.5.0"
+    "webpack": "^3.0.0"
   },
   "dependencies": {
     "classnames": "^2.2.5",
